requirements.txt
appdirs==1.4.3
appnope==0.1.0
attrs==19.3.0
awscli==1.18.29
awsebcli==3.17.1
backcall==0.1.0
bcrypt==3.1.7
blessed==1.17.4
blis==0.4.1
boto3==1.12.29
botocore==1.14.0
cached-property==1.5.1
catalogue==1.0.0
cement==2.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.3.9
cryptography==2.8
cymem==2.0.3
decorator==4.4.2
distlib==0.3.0
docker==4.2.0
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
dominate==2.5.1
filelock==3.0.12
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.16.0
guidebox==1.0.3
idna==2.7
IMDbPY==6.8
importlib-metadata==1.5.2
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.1.1
jupyter-core==4.6.3
kmodes==0.10.2
lxml==4.5.0
MarkupSafe==1.1.1
more-itertools==8.2.0
murmurhash==1.0.2
numpy==1.18.2
packaging==20.3
pandas==1.0.3
paramiko==2.7.1
parso==0.6.2
pathspec==0.5.9
pexpect==4.8.0
pickleshare==0.7.5
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
prompt-toolkit==3.0.4
psycopg2==2.8.5
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pycparser==2.20
Pygments==2.6.1
PyNaCl==1.3.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.2
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.2
pyzmq==19.0.0
requests==2.20.1
rsa==3.4.2
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
semantic-version==2.5.0
simplejson==3.17.0
six==1.11.0
sklearn==0.0
spacy==2.2.4
SQLAlchemy==1.3.16
srsly==1.0.2
termcolor==1.1.0
texttable==1.6.2
thinc==7.4.0
tmdbsimple==2.2.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
urllib3==1.24.3
virtualenv==20.0.14
visitor==0.1.3
wasabi==0.6.0
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==1.0.0
WTForms==2.2.1
zipp==3.1.0
